[daikin] Split handler class
