+++ bundles/org.openhab.binding.daikin/src/main/java/org/openhab/binding/daikin/internal/DaikinHandlerFactory.java
@@ -12,14 +12,13 @@
  */
 package org.openhab.binding.daikin.internal;
 
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.eclipse.jdt.annotation.Nullable;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.thing.ThingTypeUID;
 import org.eclipse.smarthome.core.thing.binding.BaseThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.ThingHandler;
 import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
 import org.openhab.binding.daikin.internal.handler.DaikinAcUnitHandler;
+import org.openhab.binding.daikin.internal.handler.DaikinAirbaseUnitHandler;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
@@ -32,10 +31,9 @@ import org.osgi.service.component.annotations.Reference;
 
  */
 @Component(service = ThingHandlerFactory.class, configurationPid = "binding.daikin")
-@NonNullByDefault
 public class DaikinHandlerFactory extends BaseThingHandlerFactory {
 
-    private @Nullable DaikinDynamicStateDescriptionProvider stateDescriptionProvider;
+    private DaikinDynamicStateDescriptionProvider stateDescriptionProvider;
 
     @Override
     public boolean supportsThingType(ThingTypeUID thingTypeUID) {
@@ -43,13 +41,14 @@ public class DaikinHandlerFactory extends BaseThingHandlerFactory {
     }
 
     @Override
-    protected @Nullable ThingHandler createHandler(Thing thing) {
+    protected ThingHandler createHandler(Thing thing) {
         ThingTypeUID thingTypeUID = thing.getThingTypeUID();
 
-        if (thingTypeUID.equals(DaikinBindingConstants.THING_TYPE_AC_UNIT) || thingTypeUID.equals(DaikinBindingConstants.THING_TYPE_AIRBASE_AC_UNIT)) {
+        if (thingTypeUID.equals(DaikinBindingConstants.THING_TYPE_AC_UNIT)) {
             return new DaikinAcUnitHandler(thing, stateDescriptionProvider);
+        } else if (thingTypeUID.equals(DaikinBindingConstants.THING_TYPE_AIRBASE_AC_UNIT)) {
+            return new DaikinAirbaseUnitHandler(thing, stateDescriptionProvider);
         }
-
         return null;
     }
 
@@ -58,7 +57,7 @@ public class DaikinHandlerFactory extends BaseThingHandlerFactory {
         this.stateDescriptionProvider = stateDescriptionProvider;
     }
 
-    protected  void unsetDynamicStateDescriptionProvider(DaikinDynamicStateDescriptionProvider stateDescriptionProvider) {
+    protected void unsetDynamicStateDescriptionProvider(DaikinDynamicStateDescriptionProvider stateDescriptionProvider) {
         this.stateDescriptionProvider = null;
     }
 }
